+++ src/tooling/piral-cli/src/apps/new-pilet.test.ts
@@ -2,7 +2,6 @@ import { mkdtempSync, existsSync } from 'fs';
 import { tmpdir } from 'os';
 import { join, resolve } from 'path';
 import { newPilet } from './new-pilet';
-import { SourceLanguage } from '../common';
 
 function createTempDir() {
   return mkdtempSync(join(tmpdir(), 'piral-tests-new-pilet-'));
@@ -69,7 +68,7 @@ describe('New Pilet Command', () => {
     jest.setTimeout(60000);
     const dir = createTempDir();
     await newPilet(dir, {
-      language: SourceLanguage.js,
+      language: 'js',
       install: false,
       source: 'piral@latest',
     });