+++ src/tooling/piral-cli/src/common/template.ts
@@ -1,53 +1,32 @@
-import { renderFile } from 'ejs';
-import { writeFile } from 'fs';
-import { resolve } from 'path';
-import { log } from './log';
 import { ForceOverwrite } from './enums';
 import { createFileIfNotExists } from './io';
 
-export async function applyTemplate(file: string, data: Record<string, string>) {
-  const content = await new Promise<string>((resolve, reject) => {
-    log('generalDebug_0003', `Filling template in "${file}".`);
-    renderFile(file, data, (err, str) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(str);
-      }
-    });
-  });
-  await new Promise<void>((resolve, reject) => {
-    writeFile(file, content, 'utf8', (err) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve();
-      }
-    });
-  });
+function fillTemplate(data: any = {}) {
+  return `if (process.env.NODE_ENV === 'test') {
+  // behavior for the test environment, we'll try to make it work
+
+  if (typeof window !== 'undefined') {
+    require('./${data.outFile}');
+    const ctx = window['dbg:piral'];
+    const dependencies = (ctx && ctx.pilets && ctx.pilets.getDependencies({})) || {};
+    module.exports = dependencies['${data.name}'] || {};
+  } else {
+    console.error('Your test environment does not define "window". Please make sure to provide a proper environment.');
+    module.exports = {};
+  }
+} else {
+  // under "normal" circumstances disallow such an import
+  throw new Error("This file should not be included anywhere. Usually, this means you've imported the Piral instance directly.");
+}
+`;
 }
 
-export function fillTemplate(name: string, data: any = {}) {
-  const path = resolve(__dirname, '..', '..', 'templates', `${name}.ejs`);
-  return new Promise<string>((resolve, reject) => {
-    log('generalDebug_0003', `Rendering template "${name}" in "${path}".`);
-    renderFile(path, data, (err, str) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(str);
-      }
-    });
-  });
-}
-
-export async function createFileFromTemplateIfNotExists(
-  prefix: string,
+export async function createPiralStubIndexIfNotExists(
   targetDir: string,
   fileName: string,
   forceOverwrite?: ForceOverwrite,
   data?: any,
 ) {
-  const content = await fillTemplate(`${prefix}-${fileName}`, data);
+  const content = fillTemplate(data);
   await createFileIfNotExists(targetDir, fileName, content, forceOverwrite);
 }