+++ src/tooling/piral-cli/src/helpers.test.ts
@@ -1,5 +1,5 @@
-import { keyOfForceOverwrite, keyOfSourceLanguage, valueOfForceOverwrite, valueOfSourceLanguage } from './helpers';
-import { ForceOverwrite, SourceLanguage } from './common/enums';
+import { keyOfForceOverwrite, valueOfForceOverwrite } from './helpers';
+import { ForceOverwrite } from './common/enums';
 
 describe('Piral CLI Command Helpers Module', () => {
   it('correct value of keyOfForceOverwrite results in same key', () => {
@@ -12,16 +12,6 @@ describe('Piral CLI Command Helpers Module', () => {
     expect(result).toBe('no');
   });
 
-  it('correct value of keyOfPiletLanguage results in same key', () => {
-    const result = keyOfSourceLanguage(SourceLanguage.ts);
-    expect(result).toBe('ts');
-  });
-
-  it('incorrect value of keyOfPiletLanguage results in key for TS', () => {
-    const result = keyOfSourceLanguage(5);
-    expect(result).toBe('ts');
-  });
-
   it('correct key of valueOfForceOverwrite results in same value', () => {
     const result = valueOfForceOverwrite('prompt');
     expect(result).toBe(ForceOverwrite.prompt);
@@ -30,15 +20,5 @@ describe('Piral CLI Command Helpers Module', () => {
   it('incorrect key of valueOfForceOverwrite results in first value', () => {
     const result = valueOfForceOverwrite('foo');
     expect(result).toBe(ForceOverwrite.no);
-  });
-
-  it('correct key of valueOfPiletLanguage results in same value', () => {
-    const result = valueOfSourceLanguage('ts');
-    expect(result).toBe(SourceLanguage.ts);
-  });
-
-  it('incorrect key of valueOfPiletLanguage results in value for TS', () => {
-    const result = valueOfSourceLanguage('foo');
-    expect(result).toBe(SourceLanguage.ts);
   });
 });