+++ src/tooling/piral-cli/src/types/common.ts
@@ -48,7 +48,6 @@ export interface ReleaseProvider {
 export interface TemplateFileLocation {
   from: string;
   to: string;
-  template?: boolean;
   deep?: boolean;
   once?: boolean;
 }